+++ src/gen-charts.ts
@@ -21,36 +21,29 @@ import {
 } from './core-enums.js'
 import type { IChartOptsLib, ISlideRelChart, ShadowProps, IChartPropsTitle, OptsChartGridLine, IOptsChartData, BorderProps, ChartErrorBarOptions } from './core-interfaces.js'
 import { createColorElement, createLineCap, genXmlColorSelection, genXmlPatternFill, convertRotationDegrees, encodeXmlEntities, getUuid, valToPts } from './gen-utils.js'
-import JSZip from 'jszip'
+import { ZipWriter } from './zip.js'
 
 const VALID_CHART_TIME_UNITS = ['days', 'months', 'years']
 
 /**
  * Based on passed data, creates Excel Worksheet that is used as a data source for a chart.
  * @param {ISlideRelChart} chartObject - chart object
- * @param {JSZip} zip - file that the resulting XLSX should be added to
+ * @param {ZipWriter} zip - zip writer the resulting XLSX (and chart parts) are added to
  * @return {Promise} promise of generating the XLSX file
  */
-export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JSZip): Promise<string> {
+export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: ZipWriter): Promise<string> {
 	const data = chartObject.data
 
-	return await new Promise((resolve, reject) => {
-		const zipExcel = new JSZip()
+	{
+		// The embedded workbook is its own OPC package: build it in a nested ZipWriter,
+		// then embed its bytes (no folder scaffolding; fflate emits no directory entries).
+		const zipExcel = new ZipWriter()
 		const intBubbleCols = (data.length - 1) * 2 + 1 // 1 for "X-Values", then 2 for every Y-Axis
 		const IS_MULTI_CAT_AXES = data[0]?.labels?.length > 1
 
-		// A: Add folders
-		zipExcel.folder('_rels')
-		zipExcel.folder('docProps')
-		zipExcel.folder('xl/_rels')
-		zipExcel.folder('xl/tables')
-		zipExcel.folder('xl/theme')
-		zipExcel.folder('xl/worksheets')
-		zipExcel.folder('xl/worksheets/_rels')
-
 		// B: Add core contents
 		{
-			zipExcel.file(
+			zipExcel.add(
 				'[Content_Types].xml',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?><Types xmlns="http://schemas.openxmlformats.org/package/2006/content-types">' +
 				'  <Default Extension="rels" ContentType="application/vnd.openxmlformats-package.relationships+xml"/>' +
@@ -65,7 +58,7 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 				'  <Override PartName="/docProps/app.xml" ContentType="application/vnd.openxmlformats-officedocument.extended-properties+xml"/>' +
 				'</Types>\n'
 			)
-			zipExcel.file(
+			zipExcel.add(
 				'_rels/.rels',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?><Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">' +
 				'<Relationship Id="rId1" Type="http://schemas.openxmlformats.org/package/2006/relationships/metadata/core-properties" Target="docProps/core.xml"/>' +
@@ -73,7 +66,7 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 				'<Relationship Id="rId3" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/officeDocument" Target="xl/workbook.xml"/>' +
 				'</Relationships>\n'
 			)
-			zipExcel.file(
+			zipExcel.add(
 				'docProps/app.xml',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?><Properties xmlns="http://schemas.openxmlformats.org/officeDocument/2006/extended-properties" xmlns:vt="http://schemas.openxmlformats.org/officeDocument/2006/docPropsVTypes">' +
 				'<Application>Microsoft Macintosh Excel</Application>' +
@@ -84,7 +77,7 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 				'<Company></Company><LinksUpToDate>false</LinksUpToDate><SharedDoc>false</SharedDoc><HyperlinksChanged>false</HyperlinksChanged><AppVersion>16.0300</AppVersion>' +
 				'</Properties>\n'
 			)
-			zipExcel.file(
+			zipExcel.add(
 				'docProps/core.xml',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?><cp:coreProperties xmlns:cp="http://schemas.openxmlformats.org/package/2006/metadata/core-properties" xmlns:dc="http://purl.org/dc/elements/1.1/" xmlns:dcterms="http://purl.org/dc/terms/" xmlns:dcmitype="http://purl.org/dc/dcmitype/" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">' +
 				'<dc:creator>PptxGenJS</dc:creator>' +
@@ -97,7 +90,7 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 				'</dcterms:modified>' +
 				'</cp:coreProperties>'
 			)
-			zipExcel.file(
+			zipExcel.add(
 				'xl/_rels/workbook.xml.rels',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?>' +
 				'<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">' +
@@ -107,17 +100,17 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 				'<Relationship Id="rId4" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/sharedStrings" Target="sharedStrings.xml"/>' +
 				'</Relationships>'
 			)
-			zipExcel.file(
+			zipExcel.add(
 				'xl/styles.xml',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?><styleSheet xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main"><numFmts count="1"><numFmt numFmtId="0" formatCode="General"/></numFmts><fonts count="4"><font><sz val="9"/><color indexed="8"/><name val="Geneva"/></font><font><sz val="9"/><color indexed="8"/><name val="Geneva"/></font><font><sz val="10"/><color indexed="8"/><name val="Geneva"/></font><font><sz val="18"/><color indexed="8"/>' +
 				'<name val="Arial"/></font></fonts><fills count="2"><fill><patternFill patternType="none"/></fill><fill><patternFill patternType="gray125"/></fill></fills><borders count="1"><border><left/><right/><top/><bottom/><diagonal/></border></borders><dxfs count="0"/><tableStyles count="0"/><colors><indexedColors><rgbColor rgb="ff000000"/><rgbColor rgb="ffffffff"/><rgbColor rgb="ffff0000"/><rgbColor rgb="ff00ff00"/><rgbColor rgb="ff0000ff"/>' +
 				'<rgbColor rgb="ffffff00"/><rgbColor rgb="ffff00ff"/><rgbColor rgb="ff00ffff"/><rgbColor rgb="ff000000"/><rgbColor rgb="ffffffff"/><rgbColor rgb="ff878787"/><rgbColor rgb="fff9f9f9"/></indexedColors></colors></styleSheet>\n'
 			)
-			zipExcel.file(
+			zipExcel.add(
 				'xl/theme/theme1.xml',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?><a:theme xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" name="Office Theme"><a:themeElements><a:clrScheme name="Office"><a:dk1><a:sysClr val="windowText" lastClr="000000"/></a:dk1><a:lt1><a:sysClr val="window" lastClr="FFFFFF"/></a:lt1><a:dk2><a:srgbClr val="44546A"/></a:dk2><a:lt2><a:srgbClr val="E7E6E6"/></a:lt2><a:accent1><a:srgbClr val="4472C4"/></a:accent1><a:accent2><a:srgbClr val="ED7D31"/></a:accent2><a:accent3><a:srgbClr val="A5A5A5"/></a:accent3><a:accent4><a:srgbClr val="FFC000"/></a:accent4><a:accent5><a:srgbClr val="5B9BD5"/></a:accent5><a:accent6><a:srgbClr val="70AD47"/></a:accent6><a:hlink><a:srgbClr val="0563C1"/></a:hlink><a:folHlink><a:srgbClr val="954F72"/></a:folHlink></a:clrScheme><a:fontScheme name="Office"><a:majorFont><a:latin typeface="Calibri Light" panose="020F0302020204030204"/><a:ea typeface=""/><a:cs typeface=""/><a:font script="Jpan" typeface="Yu Gothic Light"/><a:font script="Hang" typeface="맑은 고딕"/><a:font script="Hans" typeface="DengXian Light"/><a:font script="Hant" typeface="新細明體"/><a:font script="Arab" typeface="Times New Roman"/><a:font script="Hebr" typeface="Times New Roman"/><a:font script="Thai" typeface="Tahoma"/><a:font script="Ethi" typeface="Nyala"/><a:font script="Beng" typeface="Vrinda"/><a:font script="Gujr" typeface="Shruti"/><a:font script="Khmr" typeface="MoolBoran"/><a:font script="Knda" typeface="Tunga"/><a:font script="Guru" typeface="Raavi"/><a:font script="Cans" typeface="Euphemia"/><a:font script="Cher" typeface="Plantagenet Cherokee"/><a:font script="Yiii" typeface="Microsoft Yi Baiti"/><a:font script="Tibt" typeface="Microsoft Himalaya"/><a:font script="Thaa" typeface="MV Boli"/><a:font script="Deva" typeface="Mangal"/><a:font script="Telu" typeface="Gautami"/><a:font script="Taml" typeface="Latha"/><a:font script="Syrc" typeface="Estrangelo Edessa"/><a:font script="Orya" typeface="Kalinga"/><a:font script="Mlym" typeface="Kartika"/><a:font script="Laoo" typeface="DokChampa"/><a:font script="Sinh" typeface="Iskoola Pota"/><a:font script="Mong" typeface="Mongolian Baiti"/><a:font script="Viet" typeface="Times New Roman"/><a:font script="Uigh" typeface="Microsoft Uighur"/><a:font script="Geor" typeface="Sylfaen"/></a:majorFont><a:minorFont><a:latin typeface="Calibri" panose="020F0502020204030204"/><a:ea typeface=""/><a:cs typeface=""/><a:font script="Jpan" typeface="Yu Gothic"/><a:font script="Hang" typeface="맑은 고딕"/><a:font script="Hans" typeface="DengXian"/><a:font script="Hant" typeface="新細明體"/><a:font script="Arab" typeface="Arial"/><a:font script="Hebr" typeface="Arial"/><a:font script="Thai" typeface="Tahoma"/><a:font script="Ethi" typeface="Nyala"/><a:font script="Beng" typeface="Vrinda"/><a:font script="Gujr" typeface="Shruti"/><a:font script="Khmr" typeface="DaunPenh"/><a:font script="Knda" typeface="Tunga"/><a:font script="Guru" typeface="Raavi"/><a:font script="Cans" typeface="Euphemia"/><a:font script="Cher" typeface="Plantagenet Cherokee"/><a:font script="Yiii" typeface="Microsoft Yi Baiti"/><a:font script="Tibt" typeface="Microsoft Himalaya"/><a:font script="Thaa" typeface="MV Boli"/><a:font script="Deva" typeface="Mangal"/><a:font script="Telu" typeface="Gautami"/><a:font script="Taml" typeface="Latha"/><a:font script="Syrc" typeface="Estrangelo Edessa"/><a:font script="Orya" typeface="Kalinga"/><a:font script="Mlym" typeface="Kartika"/><a:font script="Laoo" typeface="DokChampa"/><a:font script="Sinh" typeface="Iskoola Pota"/><a:font script="Mong" typeface="Mongolian Baiti"/><a:font script="Viet" typeface="Arial"/><a:font script="Uigh" typeface="Microsoft Uighur"/><a:font script="Geor" typeface="Sylfaen"/></a:minorFont></a:fontScheme><a:fmtScheme name="Office"><a:fillStyleLst><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:gradFill rotWithShape="1"><a:gsLst><a:gs pos="0"><a:schemeClr val="phClr"><a:lumMod val="110000"/><a:satMod val="105000"/><a:tint val="67000"/></a:schemeClr></a:gs><a:gs pos="50000"><a:schemeClr val="phClr"><a:lumMod val="105000"/><a:satMod val="103000"/><a:tint val="73000"/></a:schemeClr></a:gs><a:gs pos="100000"><a:schemeClr val="phClr"><a:lumMod val="105000"/><a:satMod val="109000"/><a:tint val="81000"/></a:schemeClr></a:gs></a:gsLst><a:lin ang="5400000" scaled="0"/></a:gradFill><a:gradFill rotWithShape="1"><a:gsLst><a:gs pos="0"><a:schemeClr val="phClr"><a:satMod val="103000"/><a:lumMod val="102000"/><a:tint val="94000"/></a:schemeClr></a:gs><a:gs pos="50000"><a:schemeClr val="phClr"><a:satMod val="110000"/><a:lumMod val="100000"/><a:shade val="100000"/></a:schemeClr></a:gs><a:gs pos="100000"><a:schemeClr val="phClr"><a:lumMod val="99000"/><a:satMod val="120000"/><a:shade val="78000"/></a:schemeClr></a:gs></a:gsLst><a:lin ang="5400000" scaled="0"/></a:gradFill></a:fillStyleLst><a:lnStyleLst><a:ln w="6350" cap="flat" cmpd="sng" algn="ctr"><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:prstDash val="solid"/><a:miter lim="800000"/></a:ln><a:ln w="12700" cap="flat" cmpd="sng" algn="ctr"><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:prstDash val="solid"/><a:miter lim="800000"/></a:ln><a:ln w="19050" cap="flat" cmpd="sng" algn="ctr"><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:prstDash val="solid"/><a:miter lim="800000"/></a:ln></a:lnStyleLst><a:effectStyleLst><a:effectStyle><a:effectLst/></a:effectStyle><a:effectStyle><a:effectLst/></a:effectStyle><a:effectStyle><a:effectLst><a:outerShdw blurRad="57150" dist="19050" dir="5400000" algn="ctr" rotWithShape="0"><a:srgbClr val="000000"><a:alpha val="63000"/></a:srgbClr></a:outerShdw></a:effectLst></a:effectStyle></a:effectStyleLst><a:bgFillStyleLst><a:solidFill><a:schemeClr val="phClr"/></a:solidFill><a:solidFill><a:schemeClr val="phClr"><a:tint val="95000"/><a:satMod val="170000"/></a:schemeClr></a:solidFill><a:gradFill rotWithShape="1"><a:gsLst><a:gs pos="0"><a:schemeClr val="phClr"><a:tint val="93000"/><a:satMod val="150000"/><a:shade val="98000"/><a:lumMod val="102000"/></a:schemeClr></a:gs><a:gs pos="50000"><a:schemeClr val="phClr"><a:tint val="98000"/><a:satMod val="130000"/><a:shade val="90000"/><a:lumMod val="103000"/></a:schemeClr></a:gs><a:gs pos="100000"><a:schemeClr val="phClr"><a:shade val="63000"/><a:satMod val="120000"/></a:schemeClr></a:gs></a:gsLst><a:lin ang="5400000" scaled="0"/></a:gradFill></a:bgFillStyleLst></a:fmtScheme></a:themeElements><a:objectDefaults/><a:extraClrSchemeLst/><a:extLst><a:ext uri="{05A4C25C-085E-4340-85A3-A5531E510DB2}"><thm15:themeFamily xmlns:thm15="http://schemas.microsoft.com/office/thememl/2012/main" name="Office Theme" id="{62F939B6-93AF-4DB8-9C6B-D6C7DFDC589F}" vid="{4A3C46E8-61CC-4603-A589-7422A47A8E4A}"/></a:ext></a:extLst></a:theme>'
 			)
-			zipExcel.file(
+			zipExcel.add(
 				'xl/workbook.xml',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?>' +
 				'<workbook xmlns="http://schemas.openxmlformats.org/spreadsheetml/2006/main" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="x15" xmlns:x15="http://schemas.microsoft.com/office/spreadsheetml/2010/11/main">' +
@@ -128,7 +121,7 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 				'<calcPr calcId="0" concurrentCalc="0"/>' +
 				'</workbook>\n'
 			)
-			zipExcel.file(
+			zipExcel.add(
 				'xl/worksheets/_rels/sheet1.xml.rels',
 				'<?xml version="1.0" encoding="UTF-8" standalone="yes"?>' +
 				'<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">' +
@@ -193,7 +186,7 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 
 			// DONE:
 			strSharedStrings += '</sst>\n'
-			zipExcel.file('xl/sharedStrings.xml', strSharedStrings)
+			zipExcel.add('xl/sharedStrings.xml', strSharedStrings)
 		}
 
 		// tables/table1.xml
@@ -231,7 +224,7 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 			strTableXml += '</tableColumns>'
 			strTableXml += '<tableStyleInfo showFirstColumn="0" showLastColumn="0" showRowStripes="1" showColumnStripes="0"/>'
 			strTableXml += '</table>'
-			zipExcel.file('xl/tables/table1.xml', strTableXml)
+			zipExcel.add('xl/tables/table1.xml', strTableXml)
 		}
 
 		// worksheets/sheet1.xml
@@ -447,33 +440,27 @@ export async function createExcelWorksheet (chartObject: ISlideRelChart, zip: JS
 			// ....: Leaving this so nobody foolishly attempts to add this in the future
 			// strSheetXml += '<tableParts count="1"><tablePart r:id="rId1"/></tableParts>'
 			strSheetXml += '</worksheet>\n'
-			zipExcel.file('xl/worksheets/sheet1.xml', strSheetXml)
+			zipExcel.add('xl/worksheets/sheet1.xml', strSheetXml)
 		}
 
 		// C: Add XLSX to PPTX export
-		zipExcel
-			.generateAsync({ type: 'base64' })
-			.then(content => {
-				// 1: Create the embedded Excel worksheet with labels and data
-				zip.file(`ppt/embeddings/Microsoft_Excel_Worksheet${chartObject.globalId}.xlsx`, content, { base64: true })
+		// 1: Create the embedded Excel worksheet with labels and data. The xlsx is
+		//    itself a zip, so STORE it — re-DEFLATING already-compressed bytes wastes CPU.
+		zip.add(`ppt/embeddings/Microsoft_Excel_Worksheet${chartObject.globalId}.xlsx`, zipExcel.toBytes(), { store: true })
 
-				// 2: Create the chart.xml and rel files
-				zip.file(
-					'ppt/charts/_rels/' + chartObject.fileName + '.rels',
-					'<?xml version="1.0" encoding="UTF-8" standalone="yes"?>' +
-					'<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">' +
-					`<Relationship Id="rId1" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/package" Target="../embeddings/Microsoft_Excel_Worksheet${chartObject.globalId}.xlsx"/>` +
-					'</Relationships>'
-				)
-				zip.file(`ppt/charts/${chartObject.fileName}`, makeXmlCharts(chartObject))
+		// 2: Create the chart.xml and rel files
+		zip.add(
+			'ppt/charts/_rels/' + chartObject.fileName + '.rels',
+			'<?xml version="1.0" encoding="UTF-8" standalone="yes"?>' +
+			'<Relationships xmlns="http://schemas.openxmlformats.org/package/2006/relationships">' +
+			`<Relationship Id="rId1" Type="http://schemas.openxmlformats.org/officeDocument/2006/relationships/package" Target="../embeddings/Microsoft_Excel_Worksheet${chartObject.globalId}.xlsx"/>` +
+			'</Relationships>'
+		)
+		zip.add(`ppt/charts/${chartObject.fileName}`, makeXmlCharts(chartObject))
+	}
 
-				// 3: Done
-				resolve('')
-			})
-			.catch(strErr => {
-				reject(strErr)
-			})
-	})
+	// 3: Done
+	return ''
 }
 
 /**